Update setup.py and README

+++ schainpy/controller.py
@@ -47,6 +47,8 @@ def multiSchain(child, nProcess=cpu_count(), startDate=None, endDate=None, by_da
         if by_day:
             continue
         nFiles = q.get()
+        if nFiles==0:
+            continue
         firstProcess.terminate()
         skip = int(math.ceil(nFiles/nProcess))
         while True:
